changed sign up of students and companies
