+++ frontend/src/App.css
@@ -1,1 +1,11 @@
-
+h1 {
+	line-height: 1.25;
+	color: rgb(35, 33, 41);
+	font-family: Inter, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
+		"Helvetica Neue", Arial, "Noto Sans", sans-serif, "Apple Color Emoji",
+		"Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
+	font-size: 2.25rem;
+	font-weight: 800;
+	margin-bottom: 1.5rem;
+	text-align: center;
+}
