+++ frontend/src/components/Navbar/Navbar.css
@@ -18,7 +18,6 @@
 }
 
 .nav-menu {
-
 	display: flex;
 	align-items: center;
 	list-style: none;
@@ -65,6 +64,11 @@
 
 .publicLink {
 	display: flex;
+}
+
+.logo__links {
+	display: flex;
+	text-decoration: none;
 }
 
 /* .menu-icon {
